update deposit contract to be int type to comply with standard api.

+++ data/serializer/src/main/java/tech/pegasys/teku/api/response/v1/config/DepositContract.java
@@ -15,16 +15,14 @@ package tech.pegasys.teku.api.response.v1.config;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 
 public class DepositContract {
-  public final UInt64 chain_id;
+  public final int chain_id;
   public final String address;
 
   @JsonCreator
   public DepositContract(
-      @JsonProperty("chain_id") final UInt64 chain_id,
-      @JsonProperty("address") final String address) {
+      @JsonProperty("chain_id") final int chain_id, @JsonProperty("address") final String address) {
     this.chain_id = chain_id;
     this.address = address;
   }
